hamster/requirements.txt
Flask==2.2.5
google-cloud-monitoring==2.23.0
